# Optimize `get_direct_inheriters_from_class` by avoiding duplicate key lookups and value copies.